Fixing destroyed console issue
In my installation (Revo 2.3.2) this.console could not be shown because it gets an isDestroyed property.

It is quite easy to reproduce that error: Add a new package with a not existing folder two times. The second time a 'Uncaught TypeError: Cannot read property 'style' of undefined' occurs.

Maybe there is a better way to fix that issue.