Change type of Config field to mediumtext
If set to text, the maximum size of the content is just 65535 chars. When you have a package with a big config file (containing a lot of resources for example), the config gets clipped at some point, resulting in errors such as "You can not update your package because you changed package's name. Please remove package and install it again." or just empty alert messages.

Setting it to mediumtext allows for around 16Mb of content, which immediately solved the issues I was having.