Fix Field 'name' doesn't have a default value
The processor throws the following error during adding a new package:

```
Array ( [0] => HY000 [1] => 1364 [2] => Field 'name' doesn't have a default value )
```

This is because the package name is not set at that time. The whole object is saved later on, so it does not have to be saved at that time.